requirements.txt
asgiref==3.6.0
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-simple-history==3.2.0
djangorestframework==3.14.0
mysql==0.0.3
mysqlclient==2.1.1
Pillow==9.3.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
